sicoob_api/requirements.txt
fastapi
uvicorn
requests
python-dotenv
cryptography

